# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Estados",
        "description": "API para gestión de estados",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://uptyab.test/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/votacion-centros": {
            "get": {
                "tags": [
                    "VotacionCentros"
                ],
                "summary": "Obtener lista de centros de votación",
                "operationId": "f09a60b421e53dde9cf1e31364732e2b",
                "responses": {
                    "200": {
                        "description": "Lista de centros de votación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VotacionCentro"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "VotacionCentros"
                ],
                "summary": "Crear un nuevo centro de votación",
                "operationId": "599dda5b897748108c853ba6ce66a483",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ubicacion_id",
                                    "nombre",
                                    "cantidad_electores",
                                    "codigo"
                                ],
                                "properties": {
                                    "ubicacion_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "persona_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Unidad Educativa Nacional Bolivariana"
                                    },
                                    "cantidad_electores": {
                                        "type": "integer",
                                        "example": 500
                                    },
                                    "codigo": {
                                        "type": "integer",
                                        "example": 1234567890
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Centro de votación creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VotacionCentro"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/votacion-centros/{id}": {
            "get": {
                "tags": [
                    "VotacionCentros"
                ],
                "summary": "Obtener un centro de votación por ID",
                "operationId": "090e1269bb287d9df9bd261ad72fd4fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del centro de votación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Centro de votación encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VotacionCentro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Centro de votación no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "VotacionCentros"
                ],
                "summary": "Actualizar un centro de votación existente",
                "operationId": "34cdf3a035e7f30ccf2d2171057e79b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del centro de votación a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ubicacion_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "persona_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Nuevo nombre del centro"
                                    },
                                    "cantidad_electores": {
                                        "type": "integer",
                                        "example": 600
                                    },
                                    "codigo": {
                                        "type": "integer",
                                        "example": 1234567890
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Centro de votación actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VotacionCentro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Centro de votación no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "VotacionCentros"
                ],
                "summary": "Eliminar un centro de votación",
                "operationId": "6fddd3359d88880593fd6a40cc8e2042",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del centro de votación a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Centro de votación eliminado exitosamente"
                    },
                    "404": {
                        "description": "Centro de votación no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "VotacionCentro": {
                "title": "VotacionCentro",
                "description": "Modelo que representa un centro de votación",
                "properties": {
                    "id": {
                        "description": "ID del centro de votación",
                        "type": "integer",
                        "format": "int64"
                    },
                    "ubicacion_id": {
                        "description": "ID de la ubicación del centro",
                        "type": "integer",
                        "format": "int64"
                    },
                    "persona_id": {
                        "description": "ID de la persona responsable",
                        "type": "integer",
                        "format": "int64",
                        "nullable": true
                    },
                    "nombre": {
                        "description": "Nombre del centro de votación",
                        "type": "string",
                        "maxLength": 255
                    },
                    "cantidad_electores": {
                        "description": "Cantidad de electores en el centro",
                        "type": "integer"
                    },
                    "codigo": {
                        "description": "Código único del centro de votación",
                        "type": "integer",
                        "format": "int32"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}